Add files via upload
The Contact page is designed for customers to get in touch with Latto Café. It includes a simple contact form for inquiries, as well as the café’s contact details (email, phone number). There is also a map showing the café's location, allowing visitors to easily find their way. The page is styled to maintain the cozy, welcoming ambiance of the café and to make customers feel at ease while reaching out.