The Menu page
Displays the full range of offerings at Latto Café, from rich espressos to specialty coffee drinks and pastries. The page is designed with a grid layout for easy browsing, featuring images of each item and descriptions. This section also highlights the café’s seasonal specials, allowing customers to see what's currently available. The page maintains a warm and cozy aesthetic that aligns with the overall theme of the website.